Create namespaces for GraphQL types
