+++ app/graphql/types/outputs/well_type.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Types
+  module Outputs
+    # The type for Well objects.
+    class WellType < BaseObject
+      field :id, ID, null: false
+      field :created_at, String, null: false
+      field :updated_at, String, null: false
+
+      field :position, String, null: false
+      field :plate_id, Integer, null: false
+    end
+  end
+end
