--- app/graphql/types/well_mutation_type.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  # The type for Well mutations.
-  class WellMutationType < BaseObject
-    field :update_well, mutation: Mutations::UpdateWellMutation
-  end
-end
